feat: "dark mode" by inverting colors
A "proper" dark mode would be quite difficult to implement for wournal, since we
would have to somehow change the colors on the document so that contrast is
maintained despite a change in background color.

It is possible that the rest of the ui could get some prettier dark mode colors
in the future though.